Fix sign up validation
Fixes an issue that was preventing errors from being returned to the
client if requesting the ID for a user in the return result. The bug was
caused by returning a null value for this field when validation errors
occurred even though the field was marked as non-nullable in the type.